Requested changes by upstream maintainers Also rebase of minor commits. It's a learning experience...

+++ libfreerdp/core/gateway/rdg.h
@@ -141,7 +141,7 @@ struct rdp_rdg
 	int state;
 	UINT16 packetRemainingCount;
 	int timeout;
-	int extAuth;	
+	UINT16 extAuth;	
 };
